feat: tcp-echo-server uses tasks, test runner uses wasmtime cli, tcp-listener supports v6
The tcp-echo-server example has been rewritten to spawn the echo part of
each accept into a task, which  means it can accept new connections
while other echos are in flight.

The tcp-echo-server test has been rewritten to test that connections can
be accepted while other echoes are in flight.

The tcp-echo-server test has been rewritten to use wasmtime cli as a
process, rather than use wasmtime as a crate. This drops wasmtime from
the dev-dependencies of the workspace, which is good because it was
running a quite out-of-date wasmtime.

Fix up the missing conversions to/from std::net::SocketAddr in
wstd::net::tcp_listener, so that we can send a Display impl of the
listening address from the guest to host, and parse it out in the host
(see get_listening_address)