add methods for stream splice, write_all that bypasses AsyncRead/AsyncWrite traits
because the traits have ?Send bounds, but the actual AsyncInputStream
and AsyncOutputStream are both Send, so when we bypass the traits Rust
can prove that the Future for splicing is Send.

Same with write_all: when used through AsyncWrite its ?Send, but when
used on AsyncOutputStream the Future for write_all is Send.

Together, these changes allow us to show Rust that the Body::send and
http::Client::send futures are Send.