use http_body::Body (#92)
* implement bodies using http_body::Body

* erase the Incoming type

it only needs to be an internal implementation detail

* complete has trailers too

* switch up constructors

* docs, fixes, minor tweaks

* http server: more complex streaming example

* add methods for stream splice, write_all that bypasses AsyncRead/AsyncWrite traits

because the traits have ?Send bounds, but the actual AsyncInputStream
and AsyncOutputStream are both Send, so when we bypass the traits Rust
can prove that the Future for splicing is Send.

Same with write_all: when used through AsyncWrite its ?Send, but when
used on AsyncOutputStream the Future for write_all is Send.

Together, these changes allow us to show Rust that the Body::send and
http::Client::send futures are Send.

* rename forward to copy_to (code review yosh)

* remove all added async_trait stuff (unnecessary, and i dont want it)

* no you cannot